+++ integration/templates/nuxt-node/app/pages/sign-up/[...slug].vue
@@ -1,0 +1,3 @@
+<template>
+  <SignUp signInUrl="/sign-in" />
+</template>